Moved stat-related syscall handlers to new module
